requirements.txt
streamlit
python-dotenv
langchain-core
langchain-community
langchain-groq
mysql-connector-python
sounddevice
groq
deepgram-sdk